# node-lab1/api/movies/index.js
import express from 'express';
import { movies, movieReviews, movieDetails } from './moviesData.js';
import uniqid from 'uniqid'
import { notFound } from '../responses/index.js'

const router = express.Router(); 
router.get('/', (req, res) => {
    res.json(movies);
});
// Get movie reviews
router.get('/:id/reviews', (req, res) => {
    const id = parseInt(req.params.id);
    // find reviews in list
    if (movieReviews.id == id) {
        res.status(200).json(movieReviews);
    } else {
        res.status(404).json(notFound);
    }
});
// Get movie details
router.get('/:id', (req, res) => {
    const id = parseInt(req.params.id);
    if (movieDetails.id == id) {
        res.status(200).json(movieDetails);
    } else {
        res.status(404).json(notFound);
    }
});

//Post a movie review
router.post('/:id/reviews', (req, res) => {
    const id = parseInt(req.params.id);

    if (movieReviews.id == id) {
        req.body.created_at = new Date();
        req.body.updated_at = new Date();
        req.body.id = uniqid();
        movieReviews.results.push(req.body); //push the new review onto the list
        res.status(201).json(req.body);
    } else {
        res.status(404).json(notFound);
    }
});

export default router;